ApiForum. Move data-classes to shared namespace
