+++ src/back/External/ApiReport/Actions/ForumTree.php
@@ -7,8 +7,8 @@ namespace KeepersTeam\Webtlo\External\ApiReport\Actions;
 use GuzzleHttp\Exception\GuzzleException;
 use GuzzleHttp\Promise\Utils;
 use KeepersTeam\Webtlo\Data\Forum;
-use KeepersTeam\Webtlo\External\Api\V1\ForumsResponse;
 use KeepersTeam\Webtlo\External\Data\ApiError;
+use KeepersTeam\Webtlo\External\Data\ForumsResponse;
 use KeepersTeam\Webtlo\Helper;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Log\LoggerInterface;
